+++ src/main/java/by/itacademy/service/util/jwt/JwtTokenUtil.java
@@ -1,5 +1,7 @@
-package by.itacademy.service.util;
+package by.itacademy.service.util.jwt;
 
+import by.itacademy.core.dto.response.CurrentUserDto;
+import by.itacademy.core.enums.UserRole;
 import by.itacademy.service.api.IJwtTokenUtil;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -8,6 +10,9 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 @Component
@@ -21,9 +26,15 @@ public class JwtTokenUtil implements IJwtTokenUtil {
 
     @Override
     public String generateAccessToken(UserDetails user) {
+        CurrentUserDto currentUser = (CurrentUserDto) user;
+        Map<String, Object> claims = new HashMap<>();
+        claims.put("fio", currentUser.getFio());
+        claims.put("role", currentUser.getRole());
         return Jwts.builder()
-                .setSubject(user.getUsername())
+                .setSubject(currentUser.getUsername())
                 .setIssuer(properties.getIssuer())
+                .setId(currentUser.getUuid().toString())
+                .addClaims(claims)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date(System.currentTimeMillis()
                         + TimeUnit.DAYS.toMillis(7)))
@@ -40,6 +51,36 @@ public class JwtTokenUtil implements IJwtTokenUtil {
                 .getBody();
 
         return claims.getSubject();
+    }
+
+    @Override
+    public String getFio(String token) {
+        Claims claims = Jwts.parser()
+                .setSigningKey(properties.getSecret())
+                .parseClaimsJws(token)
+                .getBody();
+
+        return claims.get("fio", String.class);
+    }
+
+    @Override
+    public UUID getUuid(String token) {
+        Claims claims = Jwts.parser()
+                .setSigningKey(properties.getSecret())
+                .parseClaimsJws(token)
+                .getBody();
+
+        return claims.get("uuid", UUID.class);
+    }
+
+    @Override
+    public UserRole getUserRole(String token) {
+        Claims claims = Jwts.parser()
+                .setSigningKey(properties.getSecret())
+                .parseClaimsJws(token)
+                .getBody();
+
+        return claims.get("role", UserRole.class);
     }
 
     @Override
